Consistently use radians and degree in JoySensor
Use radians for all member variables since radians are required for the
calculations every tick. Use degree for all getters, setters and
constants as this is used by the user interface.
Use qFuzzyCompare when comparing back and forth converted floating point
numbers because they will most likely not equal to the original due to
rounding errors.

+++ src/joysensor.h
@@ -56,7 +56,7 @@ class JoySensor : public QObject
 
     Type getType();
     float getDeadZone();
-    int getDiagonalRange();
+    float getDiagonalRange();
     float getMaxZone();
     float getXCoordinate();
     float getYCoordinate();
@@ -142,7 +142,7 @@ class JoySensor : public QObject
     bool m_pending_event;
     bool m_pending_ignore_sets;
     float m_dead_zone;
-    int m_diagonal_range;
+    float m_diagonal_range;
     float m_max_zone;
     unsigned int m_sensor_delay;
